Added In-Game items
Added Item class that can be used to create weapons and items to be used by the player and enemies.

Fixed more memory leak issues regarding smart pointers (using them less frequently or at all in circular dependant classes, as it causes lingering shared_ptrs)

Player and Enemy class now inherits from a generic "CharacterBase" class to allow for items to be used on both easily.

+++ DX11Game/MenuManager.h
@@ -20,7 +20,7 @@ public:
 	_name - Unique name for menu
 	_width, _height - Resolution of the menu screen
 	*/
-	MenuNode& AddMenu(const char* _name, int _width, int _height);
+	MenuNode& AddMenu(const char* _name, float _width, float _height);
 	MenuNode& AddMenu(const char* _name, const DirectX::SimpleMath::Vector2& _dim);
 
 	// Display a menu by name
